% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_extractText.R
\name{extractText}
\alias{extractText}
\title{Read website files}
\usage{
extractText(path, ncores = 11)
}
\arguments{
\item{str}{The directory in which the files reside. Sub-directories will also be accessed}
}
\value{
data table object with two columns - file id and text
}
\description{
A Function to automatically read and convert to text the files in the given directory and sub-directories.
}
