% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_preprocessing.R
\name{preprocessingWrapper}
\alias{preprocessingWrapper}
\title{Preprocessing}
\usage{
preprocessingWrapper(texts, ids, chunk = NULL, custom_nouns = NULL)
}
\arguments{
\item{texts}{Character vector of texts}

\item{ids}{Character vector of identifiers for each text}

\item{chunk}{Boolean, whether to chunk or not. Will decide on its own if left at NULL.}

\item{custom_nouns}{Character vector of (proper) nouns to keep}
}
\value{
Quanteda tokens object
}
\description{
Wrapper for the two preprocessing steps
}
