% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_preprocessing.R
\name{preprocessWebsite}
\alias{preprocessWebsite}
\title{Preprocessing}
\usage{
preprocessWebsite(parsed, custom_nouns = NULL)
}
\arguments{
\item{parsed}{Object returned by \code{spacy_parse_chunk}}

\item{custom_nouns}{Character vector of (proper) nouns to keep}
}
\value{
Quanteda tokens object
}
\description{
Wrapper for various preprocessing steps
}
