% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_preprocessing.R
\name{spacy_parse_chunk}
\alias{spacy_parse_chunk}
\title{Chunked \code{spacy_parse}}
\usage{
spacy_parse_chunk(texts, ids, chunk = NULL)
}
\arguments{
\item{texts}{Character vector of texts}

\item{ids}{Character vector of identifiers for each text}

\item{chunk}{Boolean, whether to chunk or not. Will decide on its own if left at NULL.}
}
\value{
spacy parsed data.table
}
\description{
A wrapper for spacy_parse, but with chunking for large datasets
}
