% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_correctFiletype.R
\name{correctFiletypes}
\alias{correctFiletypes}
\title{Correct erroneous filetypes}
\usage{
correctFiletypes(path)
}
\arguments{
\item{str}{The directory in which the files reside. Sub-directories will also be accessed}
}
\description{
A Function to correct the file type of each file in the given directory. The function does not return anything - it changes the file endings on the hard drive.
}
